refactor(admin-web): 쿠폰 노출 방식 표시·입력을 어휘 정본에 붙인다 (#488 N3, P3)
