+++ apps/admin-web/src/features/mall/marketing/coupons/coupon-helpers.tsx
@@ -27,18 +27,6 @@ export function formatPeriod(coupon: MedusaPromotion): string {
   return '무기한';
 }
 
-// 판정과 트리거 어휘는 `lib/coupon-meta.ts` 로 옮겼다 — `.tsx` 는 admin-web 의 jest
-// transform(`^.+\.(t|j)s$`) 밖이라 여기 있으면 테스트가 실행되지 않는다.
-// 아래 재수출은 Task 4 에서 호출부를 새 경로로 옮긴 뒤 제거한다.
-export {
-  AUTO_ISSUE_TRIGGERS,
-  AUTO_ISSUE_TRIGGER_LABELS,
-  getCouponMeta,
-  toAutoIssueTrigger,
-  type AutoIssueTrigger,
-  type CouponMeta,
-} from './lib/coupon-meta';
-
 export function StatusBadge({ status }: { status: string }) {
   if (status === 'active') return <Badge className="bg-green-100 text-green-700 border-0">활성</Badge>;
   if (status === 'inactive') return <Badge className="bg-gray-100 text-gray-500 border-0">비활성</Badge>;
